Adapt theme to deprecated social and pagination
Using files in `layouts/partials` to override the now invalid functions
of the `blackburn` theme (author has not updated it since 2022).

Pagination changes:
- `.Page.NextPage` is now `.Page.Next`
- `.Page.PrevPage` is now `.Page.Prev`

Social changes:
`social` is no longer a top-level element in the configuration. A good
workaround is to put it in `params.social` and adapt the partials that
renders it.

I've removed some defunct / unacceptable social links from the partial
to gain a few nanoseconds when rendering.